Refactor: Move colorizers into separate file
